requirements.txt
bcrypt==4.0.1
black==23.3.0
flask==2.2.3
flask_migrate==4.0.4
flask_restful==0.3.9
flask_sqlalchemy==3.0.3
Pillow==9.5.0
pre-commit==3.2.2
psycopg2-binary==2.9.5
pyjwt==2.6.0